requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
streamlit
pandas
tqdm